+++ src/lib/pages/contract-details/components/InstantiateInfo.tsx
@@ -2,21 +2,10 @@ import { Divider, Flex, Text } from "@chakra-ui/react";
 
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { LabelText } from "lib/components/LabelText";
-import type { AddressReturnType } from "lib/hooks";
 import { useGetAddressType } from "lib/hooks";
 import type { ContractData } from "lib/model/contract";
 import { formatUTC, dateFromNow } from "lib/utils";
-
-const getAddressTypeText = (addressType: AddressReturnType) => {
-  switch (addressType) {
-    case "contract_address":
-      return "(Contract Address)";
-    case "user_address":
-      return "(Wallet Address)";
-    default:
-      return "(Invalid Address)";
-  }
-};
+import { getAddressTypeText } from "lib/utils/address";
 
 interface InstantiateInfoProps {
   contractData: ContractData;